Move neighbor search related tests to net,imglib2.neighborsearch
